Move TripTimesStringBuilder to core
