+++ src/ext-test/java/org/opentripplanner/ext/siri/SiriTimetableSnapshotSourceTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import java.time.Duration;
 import java.time.LocalDate;
 import java.time.ZoneId;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -13,7 +12,6 @@ import java.util.stream.IntStream;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.DateTimeHelper;
-import org.opentripplanner.framework.time.TimeUtils;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueStore;
 import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.calendar.CalendarServiceData;
@@ -30,6 +28,7 @@ import org.opentripplanner.transit.model.timetable.Trip;
 import org.opentripplanner.transit.model.timetable.TripOnServiceDate;
 import org.opentripplanner.transit.model.timetable.TripTimes;
 import org.opentripplanner.transit.model.timetable.TripTimesFactory;
+import org.opentripplanner.transit.model.timetable.TripTimesStringBuilder;
 import org.opentripplanner.transit.service.DefaultTransitService;
 import org.opentripplanner.transit.service.StopModel;
 import org.opentripplanner.transit.service.TransitModel;
@@ -561,7 +560,7 @@ class SiriTimetableSnapshotSourceTest {
       var tt = getTripTimesForTrip(tripId, serviceDate);
       var pattern = getPatternForTrip(tripId);
 
-      return encodeTimetable(tt, pattern);
+      return TripTimesStringBuilder.encodeTripTimes(tt, pattern);
     }
 
     public String getScheduledTimetable(String tripId) {
@@ -572,55 +571,7 @@ class SiriTimetableSnapshotSourceTest {
       var pattern = getPatternForTrip(tripId);
       var tt = pattern.getScheduledTimetable().getTripTimes(tripId);
 
-      return encodeTimetable(tt, pattern);
-    }
-
-    /**
-     * This encodes the times and information about stops in a readable way in order to simplify
-     * testing. The format is:
-     *
-     * <pre>
-     * REALTIME_STATE | stop1 [FLAGS] arrivalTime departureTime | stop2 ...
-     *
-     * Where flags are:
-     * C: Canceled
-     * R: Recorded
-     * PI: Prediction Inaccurate
-     * ND: No Data
-     * </pre>
-     */
-    private String encodeTimetable(TripTimes tripTimes, TripPattern pattern) {
-      var stops = pattern.getStops();
-
-      StringBuilder s = new StringBuilder(tripTimes.getRealTimeState().toString());
-      for (int i = 0; i < tripTimes.getNumStops(); i++) {
-        var depart = tripTimes.getDepartureTime(i);
-        var arrive = tripTimes.getArrivalTime(i);
-        var flags = new ArrayList<String>();
-        if (tripTimes.isCancelledStop(i)) {
-          flags.add("C");
-        }
-        if (tripTimes.isRecordedStop(i)) {
-          flags.add("R");
-        }
-        if (tripTimes.isPredictionInaccurate(i)) {
-          flags.add("PI");
-        }
-        if (tripTimes.isNoDataStop(i)) {
-          flags.add("ND");
-        }
-
-        s.append(" | ").append(stops.get(i).getName());
-        if (!flags.isEmpty()) {
-          s.append(" [").append(String.join(",", flags)).append("]");
-        }
-        s
-          .append(" ")
-          .append(TimeUtils.timeToStrCompact(arrive))
-          .append(" ")
-          .append(TimeUtils.timeToStrCompact(depart));
-      }
-      return s.toString();
+      return TripTimesStringBuilder.encodeTripTimes(tt, pattern);
     }
 
     /**
